Refactor provider into flat structure
