+++ main.go
@@ -3,13 +3,13 @@ package main
 import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/plugin"
-	"github.com/symopsio/terraform-provider-sym/sym"
+	"github.com/symopsio/terraform-provider-sym/sym/provider"
 )
 
 func main() {
 	plugin.Serve(&plugin.ServeOpts{
 		ProviderFunc: func() *schema.Provider {
-			return sym.Provider()
+			return provider.Provider()
 		},
 	})
 }
