+++ sym/provider/util.go
@@ -1,4 +1,4 @@
-package resources
+package provider
 
 import (
 	"fmt"
